Make conflicted documents not crash on find/reload
The first step to handling conflicts is to fix document attribute
loading to not crash hard when conflicts are present.
Riak::Robject handles this gracefully, whereas Riak::Document
tries to call hash methods on a string.

This commit makes documents not attempt to assign attributes
in the case of conflict. It would now be *possible* to build a
system with conflict resolution using ripple, although if a
developer were to set Bucket#allow_mult = true and not handle a
documents' robjects' conflicts, undefined behavior would
definitely result in data loss on save.

Making a better API for conflict handling is forthcoming.

Working towards closing #24.